Move sorting fetch tokens functions
