+++ packages/synapse-interface/utils/actions/fetchPortfolioBalances.tsx
@@ -153,3 +153,38 @@ export const fetchPortfolioBalances = async (
     return { balancesAndAllowances: {}, status: FetchState.INVALID, error }
   }
 }
+
+export function separateTokensByAllowance(
+  tokens: TokenWithBalanceAndAllowances[]
+): [TokenWithBalanceAndAllowances[], TokenWithBalanceAndAllowances[]] {
+  const tokensWithAllowance: TokenWithBalanceAndAllowances[] = []
+  const tokensWithoutAllowance: TokenWithBalanceAndAllowances[] = []
+
+  tokens &&
+    tokens.forEach((token: TokenWithBalanceAndAllowances) => {
+      // currently separating by bridge allowance
+      // update this when incorporating other allowances to order by
+      const bridgeAllowance: bigint | null = token.allowances[ROUTER_ADDRESS]
+      if (bridgeAllowance === null) {
+        tokensWithAllowance.push(token)
+      } else if (bridgeAllowance > 0n) {
+        tokensWithAllowance.push(token)
+      } else {
+        tokensWithoutAllowance.push(token)
+      }
+    })
+
+  return [tokensWithAllowance, tokensWithoutAllowance]
+}
+
+export function sortByBalanceDescending(
+  tokens: TokenWithBalanceAndAllowances[]
+): TokenWithBalanceAndAllowances[] {
+  return (
+    tokens &&
+    tokens.sort(
+      (a: TokenWithBalanceAndAllowances, b: TokenWithBalanceAndAllowances) =>
+        b.parsedBalance > a.parsedBalance ? 1 : -1
+    )
+  )
+}
